requirements.txt
cffi==1.14.3
faiss-cpu==1.6.3
future==0.18.2
numpy==1.19.2
opencv-python==4.4.0.42
pandas==1.1.2
Pillow==7.2.0
pycparser==2.20
pyheif==0.5.1
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
torch==1.6.0
torchvision==0.7.0
